% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_.R
\name{calc_sainsbury}
\alias{calc_sainsbury}
\title{Calcula el reach usando el modelo de Sainsbury}
\usage{
calc_sainsbury(audiencias, poblacion_total)
}
\arguments{
\item{audiencias}{Vector numérico con las audiencias de cada soporte}

\item{poblacion_total}{Número total de la población objetivo}
}
\value{
Una lista con los siguientes elementos:
\itemize{
\item reach$porcentaje: Reach total en porcentaje
\item reach$personas: Reach total en número de personas
\item distribucion$porcentaje: Vector con la distribución de contactos en porcentaje
\item distribucion$personas: Vector con la distribución de contactos en personas
\item acumulada$porcentaje: Vector con la distribución acumulada en porcentaje
\item acumulada$personas: Vector con la distribución acumulada en personas
}
}
\description{
Calcula el reach usando el modelo de Sainsbury
}
\examples{
audiencias <- c(300000, 400000, 200000)
poblacion_total <- 1000000
resultado <- calc_sainsbury(audiencias, poblacion_total)
}
