% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reach_frequency.R
\name{calcular_R1_R2}
\alias{calcular_R1_R2}
\title{Calcula R1 y R2 a partir de los par치metros A y B}
\usage{
calcular_R1_R2(A, B)
}
\arguments{
\item{A}{Numeric. Valor del par치metro A.}

\item{B}{Numeric. Valor del par치metro B.}
}
\value{
Una lista con los valores de R1 y R2.
}
\description{
Calcula R1 y R2 a partir de los par치metros A y B
}
\examples{
calcular_R1_R2(A = 0.5, B = 0.3)
}
