% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_canex.R
\encoding{UTF-8}
\name{calculate_duplication}
\alias{calculate_duplication}
\title{Calculate Vehicle Duplication}
\usage{
calculate_duplication(pij, pi, pj)
}
\arguments{
\item{pij}{Numeric. Joint exposure probability}

\item{pi}{Numeric. First vehicle marginal probability}

\item{pj}{Numeric. Second vehicle marginal probability}
}
\value{
Numeric. Correlation coefficient (-1 to 1)
}
\description{
Calculates correlation coefficient between two vehicles
based on their reach probabilities.
}
\examples{
corr <- calculate_duplication(0.15, 0.3, 0.4)

}
