% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaPlanR-package.R
\docType{package}
\name{mediaPlanR}
\alias{mediaPlanR}
\alias{mediaPlanR-package}
\title{Herramientas para Planificaci칩n de Medios}
\description{
Conjunto de funciones y aplicaciones para planificaci칩n de medios.
}
\details{
mediaPlanR: Herramientas para Planificaci칩n de Medios
}
\references{
\href{https://cran.r-project.org/package=mediaPlanR}{CRAN}
}
\author{
Manuel S치nchez-Franco \email{majesus@us.es}
}
