% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reach_frequency.R
\name{optimizar_y_calcular}
\alias{optimizar_y_calcular}
\title{Optimiza la distribución de contactos y calcula R1 y R2}
\usage{
optimizar_y_calcular(POB, Pi, valor_objetivo, salto_A = 0.025, salto_B = 0.025)
}
\arguments{
\item{POB}{Numeric. Tamaño de la población objetivo.}

\item{Pi}{Numeric. Valor objetivo de distribución de contactos Pi.}

\item{valor_objetivo}{Numeric. Número de personas a alcanzar Pi veces.}

\item{salto_A}{Numeric. Paso para el rango de probabilidad alpha.}

\item{salto_B}{Numeric. Paso para el rango de probabilidad beta.}
}
\value{
Data frame con las combinaciones óptimas de x, alpha, R1, R2, y prob.
}
\description{
Esta función optimiza la distribución de contactos y calcula los valores de R1 y R2
en función de los parámetros proporcionados.
}
\examples{
optimizar_y_calcular(POB = 1000000, Pi = 3, valor_objetivo = 0.043, salto_A = 0.125, salto_B = 0.125)
}
