% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_MBBD.R
\name{print.MBBD}
\alias{print.MBBD}
\title{Imprime un objeto MBBD}
\usage{
\method{print}{MBBD}(x, ...)
}
\arguments{
\item{x}{Objeto de clase "MBBD"}

\item{...}{Argumentos adicionales pasados a print}
}
\description{
Método para imprimir los resultados de un modelo MBBD
}
