% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reach_frequency.R
\name{reach_frequency}
\alias{reach_frequency}
\title{Calcula el alcance y la frecuencia}
\usage{
reach_frequency(impresiones, audiencia)
}
\arguments{
\item{impresiones}{Número total de impresiones de la campaña}

\item{audiencia}{Tamaño total de la audiencia objetivo}
}
\value{
Una lista con el alcance (en porcentaje) y la frecuencia media
}
\description{
Esta función calcula el alcance y la frecuencia de una campaña publicitaria
basándose en el número de impresiones y el tamaño de la audiencia.
}
\examples{
reach_frequency(1000000, 500000)
}
