% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_binomial_explorer.R
\encoding{UTF-8}
\name{run_beta_binomial_explorer}
\alias{run_beta_binomial_explorer}
\title{Función de Masa de Probabilidad Beta-Binomial}
\usage{
run_beta_binomial_explorer()
}
\value{
Una aplicación Shiny
}
\description{
Ejecuta la aplicación Shiny del explorador Beta Binomial
}
