% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{run_mediaPlanR}
\alias{run_mediaPlanR}
\title{Ejecutar Aplicaciones de MediaPlanR}
\usage{
run_mediaPlanR(app = c("beta", "aud", "reach"))
}
\arguments{
\item{app}{character: Nombre de la aplicación a ejecutar ("beta", "aud", o "reach")}
}
\description{
Ejecutar Aplicaciones de MediaPlanR
}
