% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reach_converg_explorer.R
\encoding{UTF-8}
\name{run_reach_converg_explorer}
\alias{run_reach_converg_explorer}
\title{Explorador de Convergencia de la Cobertura}
\usage{
run_reach_converg_explorer()
}
\description{
Aplicación Shiny para el análisis de la convergencia de la cobertura.
}
\details{
La aplicación permite:
\itemize{
\item Configurar un plan de medios aplicando el modelo Beta-Binomial
\item Analizar la evolución de la cobertura acumulada e incremental
\item Analizar la distribución de contactos (y acumulada)
\item Visualizar distribuciones mediante gráficos de líneas
\item Calcular estadísticas relevantes de la audiencia
}
}
\section{Parámetros de Configuración}{

\itemize{
\item Tamaño de población
\item Parámetros de forma de la distribución Beta-Binomial
\item Máximo número de contactos a mostrar
\item Umbral de convergencia
}
}

