% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_binomial_explorer.R
\docType{data}
\name{ui}
\alias{ui}
\title{Interfaz de usuario para la aplicación Shiny del explorador Beta Binomial}
\format{
An object of class \code{bslib_page} (inherits from \code{shiny.tag.list}, \code{list}) of length 5.
}
\usage{
ui
}
\value{
Una interfaz de usuario Shiny
}
\description{
Interfaz de usuario para la aplicación Shiny del explorador Beta Binomial
}
\details{
UI Function
}
\keyword{datasets}
