% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_.R
\name{calc_beta_binomial}
\alias{calc_beta_binomial}
\title{Calcula el reach usando el modelo Beta-Binomial}
\usage{
calc_beta_binomial(A1, A2, P, n)
}
\arguments{
\item{A1}{Numeric. Audiencia tras la primera inserción}

\item{A2}{Numeric. Audiencia tras la segunda inserción}

\item{P}{Numeric. Población total objetivo}

\item{n}{Integer. Número de inserciones totales}
}
\value{
Una lista con los siguientes elementos:
\itemize{
\item reach$porcentaje: Reach total en porcentaje
\item reach$personas: Reach total en número de personas
\item distribucion$porcentaje: Vector con la distribución de contactos en porcentaje
\item distribucion$personas: Vector con la distribución de contactos en personas
\item acumulada$porcentaje: Vector con la distribución acumulada en porcentaje
\item acumulada$personas: Vector con la distribución acumulada en personas
\item parametros: Lista con parámetros alpha, beta y probabilidad de cero contactos
}
}
\description{
Calcula el reach usando el modelo Beta-Binomial
}
\examples{
resultado <- calc_beta_binomial(
  A1 = 500000,
  A2 = 550000,
  P = 1000000,
  n = 5
)
}
