% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reach_frequency.R
\name{calcular_R1_R2}
\alias{calcular_R1_R2}
\title{R1 y R2 a partir de parámetros alpha y beta}
\usage{
calcular_R1_R2(A, B)
}
\arguments{
\item{A}{Parámetro alpha, debe ser numérico y positivo}

\item{B}{Parámetro beta, debe ser numérico y positivo}
}
\value{
Lista con valores R1 y R2 calculados
}
\description{
Esta función calcula R1 y R2 a partir de los parámetros alpha y beta
}
\examples{
calcular_R1_R2(0.5, 0.3)
}
