% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reach_frequency.R
\name{imprimir_resultados}
\alias{imprimir_resultados}
\title{Impresión de resultados del análisis}
\usage{
imprimir_resultados(data_ls)
}
\arguments{
\item{data_ls}{Una lista que contiene:
\itemize{
\item resultados - Data frame con resultados
\item distribucion - Data frame con la distribución
\item alpha - Valor del parámetro alpha
\item beta - Valor del parámetro beta
}}
}
\value{
No retorna valor, imprime los resultados en consola
}
\description{
Esta función imprime en consola los resultados del análisis incluyendo
combinaciones, distribución y parámetros seleccionados.
}
\details{
Imprime resultados del análisis
}
\examples{
data_ls <- list(
  resultados = data.frame(x = 1:3, y = 4:6),
  distribucion = data.frame(cont = 1:3, prob = c(0.3, 0.4, 0.3)),
  alpha = 0.5,
  beta = 0.3
)
\dontrun{
imprimir_resultados(data_ls)
}

}
