% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_binomial_explorer.R
\encoding{UTF-8}
\name{run_beta_binomial_explorer}
\alias{run_beta_binomial_explorer}
\title{Explorador de la Distribución Beta-Binomial}
\usage{
run_beta_binomial_explorer()
}
\description{
Aplicación Shiny que permite explorar y visualizar la
distribución beta-binomial de manera interactiva.
}
