% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_binomial_explorer.R
\docType{data}
\encoding{UTF-8}
\name{ui}
\alias{ui}
\title{Explorador Interactivo de la Distribución Beta-Binomial}
\format{
An object of class \code{bslib_page} (inherits from \code{shiny.tag.list}, \code{list}) of length 5.
}
\usage{
ui
}
\description{
Aplicación Shiny que permite explorar y visualizar la
distribución beta-binomial de manera interactiva.
}
\examples{
\dontrun{
  run_beta_binomial_explorer()
}

}
\keyword{datasets}
